initscripts: banner.sh: Prefer /dev/console to /dev/tty0
Prefer writing the banner to /dev/console instead of /dev/tty0. In many
cases, /dev/console is an alias for /dev/tty0, but sometimes it is not
(if for example the default console is changed on the kernel command
line). Writing the banner to /dev/tty0 can have some undesirable side
effects if /dev/console has been redirected, such as writing on top of a
kernel splash screen being displayed as a boot logo.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>