kernel-dev: introduce optional packages linux-headers and linux-source
Creating two new packages to go along with the minimal kernel-devsrc
package that is used for building out of tree modules against the
running kernel.

These two packages are for compatibility or use cases that
need entire source code or headers (and things we don't necessarily want
in the minimal devsrc package)

 linux-source package installs to: /usr/src/linux-source-<version>
 linux-headers package install to: /usr/src/linux-headers-<version>

As a comparison, ubuntu installs both of the above:

   ii  linux-headers-4.18.0-17      4.18.0-17.18   all Header files related to Linux kernel version 4.18.0
   ii  linux-source-4.18.0          4.18.0-20.21   all Linux kernel source for version 4.18.0 with Ubuntu patches

When kernel-devsrc, linux-headers and linux-source are installed to an
image, the result is:

   root@qemux86-64:/usr/src# ls -alF
      total 16
      drwxr-xr-x  4 root root 4096 May 28 15:52 ./
      drwxr-xr-x 10 root root 4096 May 28 13:38 ../
      lrwxrwxrwx  1 root root   46 May 28 13:38 kernel -> ../../lib/modules/5.0.17-yocto-standard/source/
      lrwxrwxrwx  1 root root   35 May 28 15:51 linux-headers -> linux-headers-5.0.17-yocto-standard/
      drwxr-xr-x 23 root root 4096 May 28 15:49 linux-headers-5.0.17-yocto-standard/
      lrwxrwxrwx  1 root root   34 May 28 15:52 linux-source -> linux-source-5.0.17-yocto-standard/
      drwxr-xr-x  2 root root 4096 May 28 15:49 linux-source-5.0.17-yocto-standard/

   root@qemux86-64:/usr/src# du -shL kernel
      89M     kernel
   root@qemux86-64:/usr/src# du -shL linux-headers
      114M    linux-headers
   root@qemux86-64:/usr/src# du -shL linux-source
      123M    linux-source
   root@qemux86-64:/usr/src# ls -alF linux-source/
      total 123M
      drwxr-xr-x 2 root root      4.0K May 28 15:49 ./
      drwxr-xr-x 4 root root      4.0K May 28 15:52 ../
      -rw-r--r-- 1 root root 123M May 28 13:41 linux-source-5.0.17-yocto-standard.tar.bz2

   root@qemux86-64:/usr/src# rpm -qa | grep linux-headers
      linux-headers-5.0.17-yocto-standard-1.0-r0.qemux86_64
   root@qemux86-64:/usr/src# rpm -qa | grep linux-source
      linux-source-5.0.17-yocto-standard-1.0-r0.qemux86_64

Note: when uncompressed the linux-source tarball is about 910M

Since the linux-source copy takes a long time, it is optional via the
"linux-source" PACKAGECONFIG variable. Additionally, the full source can
be compressed or uncompressed and is controlled via the KERNELDEV_COMPRESSED_SOURCE
variable.

The generic symlinks created to the versioned linux-source and headers
are done via postinst added through anonymous python. This technique is
due to issues some package managers that are throwing errors when a
postinst is added to a package that is renamed via PGK_ mappings.
With the python technique (versus the commented out blocks
below), we can give full control on whether or not the symlinks are
created AND if the postinsts are even created.

The variable KERNELDEV_SYMLINKS controls whether or not the symlinks
are created.

We are not using update-alternatives, since it has caused problems
for kernel packages in the past ([kernel: Stop using update-alternatives]) due
to relative symlinks. We are also not installing the symlink as part of the
package, since that breaks upgrade scenarios where two packages need to be
installed at the same time (and we have a file conflict).

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>