binconfig: cleanup more paths in config scripts
If a config file invokes another config file then there is a chance that it uses
the absolute path of the config file (for example, xslt-config invoking
xml-config).  To fix the paths swap STAGING_BINDIR_CROSS with ${bindir}.

DEBUG_PREFIX_MAP and TOOLCHAIN_OPTIONS can often work their way into config
scripts which we don't want, so strip those out entirely.

Fix up some replacements which were not explicitly global.

Finally, add a bbdebug so this operation is in the debug log.

Signed-off-by: Ross Burton <ross.burton@intel.com>