binconfig: unify behaviour with binconfig-disabled
Unify the behaviour of this class with binconfig-disabled which has far more
sensible behaviour.  Currently BINCONFIG_GLOB is a filename glob (defaulting to
*-config) which mangles all *installed* files in the sysroot but files found in
${S} and ${B} for the packages (*not* the installed files in ${D}, which causes
differences if do_install modifies the files).  Also FILES_${PN}-dev defaults to
${bindir}/*-config, but the glob matches files in any directory.

The new semantics, matching binconfig-disabled, is that BINCONFIG is a
FILES-style list of globs, defaulting to ${bindir}/*-config.  Only files that
are installed to ${D} are processed, not anywhere else.

To help migration to this new model, if BINCONFIG_GLOB is set in a recipe an
error is emitted.

Signed-off-by: Ross Burton <ross.burton@intel.com>