binconfig-disabled: don't disable packaged scripts
Whilst disabling the binconfig scripts in the sysroot is typically sensible as
these scripts are often not friendly to cross-compilation or sysroots, they tend
to work on the target and more importantly we don't control what is being built
by the user on the target.

Change binconfig-disabled.bbclass to use the logic in binconfig.bbclass to fixup
the packaged binconfig scripts and leave them in the packages, whilst still
disabling the sysroot binconfig scripts.

Signed-off-by: Ross Burton <ross.burton@intel.com>