oeqa/kerneldevelopment: Able to apply a single patch to the Linux kernel source
Purpose of kernel development is basically to customize our own
recipes kernel by reused existing recipes.

This is an initiative of automating manual kernel development test
cases. Applying a single patch to the Linux kernel source is one
of the manual test cases of kernel development. Purpose of this
test is as a developer we can make changes of a file in kernel
source and able to apply a single patch to the kernel source.

ref:  https://wiki.yoctoproject.org/wiki/Kernel_Development_Test_Cases

Signed-off-by: Mazliana <mazliana.mohamad@intel.com>