scripts/create-pull-request: improve handling for
When attempting to create a pull request we look at the remote URL in
order to extract information to include in the cover letter. However,
the assumption was that the remote is an SSH URL that the user could
push to, which is not always the case (the pull and push URLs might be
different, or we might be pushing via https) - if it wasn't the script
just gave up leaving you to manually edit the URLs in the email. With
a few minor tweaks to the regexes the script will work for these cases
as well.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>