buildhistory: get rid of BUILDHISTORY_FEATURES checks in feature classes themselves; automatically inherit parent class
If someone wants to subclass one of the feature classes, e.g.
buildhistory-package, they may expect to be able to just inherit that
class directly. This will let that work.

Of course, then they will also get other features declared in
BUILDHISTORY_FEATURES automatically. But that is the same
behavior as when buildhistory was a single class anyway, so it should be
expected. I will be sure to document it regardless.

Signed-off-by: Chris Laplante <chris.laplante@agilent.com>