mesa: Upgrade to 19.2.4
Upgrade mesa and mesa-gl to 19.2.4.

The license hash change was a trivial new line removal.

The glx-tls option was removed as it isn't included in the meson.build
file. It has been replaced with 'use-elf-tls' instead.

I have backported the asm removal as an attempt to fix the musl build
issue.

(From OE-Core rev: 42671204ede483873379ecbf36b80f9a35118d21)

Signed-off-by: Alistair Francis <alistair@alistair23.me>
Signed-off-by: Alistair Francis <alistair.francis@wdc.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>