bitbake: EXPERIMENT: hashserv: Allow equiv mappings
Experiment with allowing equivalent hashes to be mapped to each other.

Adds the server extension and updates the code to call into it when
the scenario is detected.

(gdb-cross dependency of meta-extsdk-toolchain which can be x86 or arm
yet the downstream hash needs to match)

(Bitbake rev: 4868b089fe5a8c56246b55dbfb312a34a3a5e44a)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>