selftest-hardlink: don't RDEPEND on gdb explicitly
For easy of reuse and speed of building, don't RDEPEND on gdb
explicitly.  Instead put gdb.sh into ${PN} and explicitly install gdb
into the image if needed.

This means other tests that don't need to use gdb can reuse
this recipe without waiting for a target gdb to build.

Signed-off-by: Ross Burton <ross.burton@intel.com>