classes/reproducible_build: Only whitlist SDE if not fixed
If the SOURCE_DATE_EPOCH variable is a fixed integer value, it should be
included in taskhashes (not present in BB_HASHBASE_WHITELIST). In all
other cases, it must be excluded from taskhashes because it depends on
information provided in do_unpack and cannot be calculated correctly at
parsing time.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>