kernel: provide alternate headers for package specific interfaces
Provide alternate uapi headers from a kernel, for use by
a recipe. This is *not* a replacement for the linux-libc-headers
package, but is expected to be used when a package is tightly
coupled to a specific kernel and needs definitions / interfaces
specific to that kernel.

The class works in one of two ways, it can be included by the
recipe that requires the headers (mode 1 below) or by a kernel
recipe that provides the alternate headers. Both are valid,
and one may be chosen over the other depending on how many
packages need custom headers, etc.

It is up to the package that builds against these headers to
ensure that the alternate kernel headers are in the include
path and used for the build.

mode 1: included directly from a recipe that needs alternate
        kernel headers. In this mode the headers install from
        the shared kernel dir to WORKDIR.
        Note: in this mode, the kernel doesn't (or shouldn't)
              inherit kernel-alt-headers itself.

        The headers will be in: $WORKDIR/usr/alt-headers

mode 2: included by a kernel recipe to provide alt-headers.
        In this mode, the kernel build/install is modified to
        stage the headers to the sysroot (and create a
        kernel-headers-alt package).

        The headers are found in: recipe-sysroot/usr/alt-headers

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>