fetch2/gitsm: Make need_update() process submodules
If the bitbake.srcrev nugget is not present for the commit we're
interested in we should not just bail out and say that an update is
needed. Instead we can recursively walk through the submodules and check
for the presence of the required commits.

Signed-off-by: Paul Barker <pbarker@konsulko.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit cfc78316309556bec487ef0a5a9205e41f1be86f)
Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>