sstate: Track sstate status persistently
Hash equivalence causes the runqueue to switch between executing regular
tasks and executing setscene tasks, and each time it must recheck for
valid sstate objects. In most common setups, the sstate cache is not
going to contain anything new since the last time a check was made, so
instead of always re-searching for items, make set of found and missing
items persistent. This provides a nice positive and negative cache so
that the items

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>