rt-tests: Update recipes to use 1.8
John Kacur(the maintainer of rt-tests), recommends "that they use
unstable/devel/latest as the maintained stable branch". I received this
information on IRC, I couldn't find it documented anywhere else.

This means that both rt-tests and hwlatdetect can be updated to be 1.8,
the latest release on unstable/devel/latest.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>