librsvg: Update from 2.40.20 to 2.46.4
The new version is written in Rust.
Remove the patches that were backported and drop the reproducibility
patch since that code was dropped in the Rust translation.

Update the license checksums since the license info was updated
to be more consistent in commit:
bd8ebc43 Use the LGPL 2.1 consistently
Fixes these:
   https://gitlab.gnome.org/GNOME/librsvg/issues/72
   https://gitlab.gnome.org/GNOME/librsvg/issues/432

Also inherit gettext to address:

Maybe:
***
QA Issue: AM_GNU_GETTEXT used but no inherit gettext
and sort the inherit list for good measure.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>