qemuboot: make default for QB_DEFAULT_KERNEL a weak assignment
When attemping Xen qemu boots out of meta-virt, it was discovered
that the deafult in the xen images weren't correct for all use
cases.

These assignents were done with QB_DEFAULT_KERNEL = "", etc.

To allow overriding of those values in specific implementations,
we change the meta-virt ones to ?= assignments.

Following the chain of variable manipulations, if we make the
default ??=, then the layers can all set their value as they
expect.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>