package.bbclass: fix image_debugfs symlinks
Allow symlinks to be made in the debugfs by removing an incorrect check that
would always prevent symlinks from being transferred over.

The comment in this code indicates that the code is meant to ensure symlinks
are only copied to the debugfs if their target has already been copied, but
instead checks for the existence of the symlink that has yet to be created.
As a result this check always fails.

This check is safe to get rid of since we will only attempt to copy a
symlink to the debugfs if it's original target exists and is an ELF. It
is, however, still possible for a dead symlink to end up in the debugfs
if its original target was stripped.

Signed-off-by: William Hauser <william.hauser@meraki.com>