kernel: provide virtual/kernel for all kernel providers
Since we allow the kernel package name to be varied, we should
allow those packages to provide virtua/kernel, so they can be used
as primary kernels.

This isn't using a regex, since that would further (and in a hard
to find way) limit what is or isn't a provider.

There are currently no recipes that inherit kernel that are not
kernel recipes (kernel-base, etc, provide other facilities), so
making this provides simpler shouldn't case problems.

Multiple kernel providers can be deal with in the same way as
any multiple provider.

[YOCTO: #13172]

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>